+++ packages/spacecat-shared-http-utils/CHANGELOG.md
@@ -1,6 +1,25 @@
+# [@adobe/spacecat-shared-http-utils-v1.1.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-http-utils-v1.1.0...@adobe/spacecat-shared-http-utils-v1.1.1) (2024-01-13)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#81](https://github.com/adobe/spacecat-shared/issues/81)) ([67cf21c](https://github.com/adobe/spacecat-shared/commit/67cf21c548d036de45221e118ff3bc0e7b26a692))
+
+# [@adobe/spacecat-shared-http-utils-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-http-utils-v1.0.0...@adobe/spacecat-shared-http-utils-v1.1.0) (2024-01-12)
+
+
+### Bug Fixes
+
+* github org ([02f86c3](https://github.com/adobe/spacecat-shared/commit/02f86c351b2b1ab99ce5aca7fd5e6af2246c0efd))
+
+
+### Features
+
+* site delivery type ([#77](https://github.com/adobe/spacecat-shared/issues/77)) ([005bc38](https://github.com/adobe/spacecat-shared/commit/005bc388ed6ab0f82c8b0f562d4c1f4c0d1e2894))
+
 # @adobe/spacecat-shared-http-utils-v1.0.0 (2023-12-14)
 
 
 ### Features
 
-* http utils package ([#51](https://github.com/adobe-rnd/spacecat-shared/issues/51)) ([84cd91a](https://github.com/adobe-rnd/spacecat-shared/commit/84cd91ae4225455327116169f5b040cc9adc1c2e))
+* http utils package ([#51](https://github.com/adobe/spacecat-shared/issues/51)) ([84cd91a](https://github.com/adobe/spacecat-shared/commit/84cd91ae4225455327116169f5b040cc9adc1c2e))